pyproject.toml
[tool.poetry]
name = "ics"
version = "0.8.0-dev"
description = "Pythonic iCalendar (RFC 5545) Parser"
authors = ["Nikita Marchant <nikita.marchant@gmail.com>", "Niko Fink <icspy@niko.fink.bayern>"]
license = "Apache-2.0"
readme = "README.rst"
homepage = "https://pypi.org/project/ics/"
repository = "https://github.com/C4ptainCrunch/ics.py"
documentation = "https://icspy.readthedocs.io/en/stable/"
keywords = ["ics", "icalendar", "calendar", "event", "rfc5545"]
classifiers = [
    'Development Status :: 4 - Beta',
    'Topic :: Software Development :: Libraries',
    'Topic :: Software Development :: Libraries :: Python Modules',
    'Intended Audience :: Developers',
    'Topic :: Office/Business :: Scheduling',
    'License :: OSI Approved :: Apache Software License',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3 :: Only',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Typing :: Typed',
]

[tool.poetry.dependencies]
python =                     "^3.6"
python-dateutil =            "^2.8"
attrs =                      ">=19.2"
tatsu =                      ">4.2"
importlib_resources =        "^1.4"
ics_vtimezones =             "^2020.1"

# extra: test
pytest =                     { version = "^5.2",     optional = true }
pytest-cov =                 { version = "^2.8.1",   optional = true }
hypothesis =                 { version = "^5.8.0",   optional = true }

# extra: dev
bump2version =               { version = "^1.0.0",   optional = true }
tox =                        { version =  "^3.15.0", optional = true }

[tool.poetry.extras]
test = ["pytest", "pytest-cov", "hypothesis"]
dev = ["bump2version", "tox"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
